new changes

+++ database/migrations/2023_12_01_082457_create_campaign_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('company_id');
             $table->string('title')->nullable();
             $table->text('description')->nullable();
-            $table->float('reward')->nullable();
+            $table->integer('reward')->nullable();
             $table->string('image')->nullable();
             $table->date('expiry_date');
             $table->enum('type', ['1', '2' ,'3'])->default('1')->comment('1) Referral 2) Social 3) Custom');
